Splinting init.pp, updating and extend example, update import_templates, update README
Split init.pp into install.pp, config.pp and service.pp.

Move graph creation to implementation since leaving it in node.pp the
way it was before caused problems with other nodes. (The way I had
implemented it in node.pp was because I did not understand how include
worked in puppet)

Templates are now only imported on the first fun and re-imported when
the template files change.